Moved functions applying overrides to a new file.
